requirements.txt
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg[binary]==3.2.9
pillow==11.3.0
django-cors-headers==4.7.0
python-decouple==3.8
python-decouple==3.8
channels==4.3.1
celery==5.5.3
whitenoise==6.10.0
eventlet==0.40.2
redis==6.4.0